Remove legacy contextual dialog factories and views
